\name{hgu133plus2barcodevecs}
\docType{data}
\alias{hgu133plus2barcodevecs}
\title{hgu133plus2 data for barcode}
\description{Data used by the barcode package for microarrays of type hgu133plus2.}
\usage{
  data(bcparams-GPL570)
}
\format{
  A matrix of parameters estimates used in the barcode function for
  Affymetrix hgu133plus2 microarray data.
  }
\keyword{datasets}


